pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shadow-gentcore"
version = "0.1.0"
description = "Multi-domain AI agent framework SDK — runtime, CLI, authoring tools, and orchestration engine."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "agent-contracts>=0.1.0",
    "agent-tools>=0.1.0",
    "anthropic>=0.30",
    "httpx>=0.27",
    "pydantic>=2.0,<3.0",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]
all-providers = [
    "openai>=1.0",
    "boto3>=1.34",
]
server = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
]

[project.scripts]
ai = "harness.cli.ai:cli"

[tool.hatch.build.targets.wheel]
packages = ["harness", "agents", "workflows", "config"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["harness/tests"]
asyncio_mode = "auto"
